# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Loop & Love API",
    "version": "1.0.0",
    "description": "Contrato OpenAPI de la API REST actualmente implementada por Loop & Love. La especificación refleja únicamente las rutas y comportamientos existentes en el backend."
  },
  "servers": [
    {
      "url": "https://loop-and-love-y5gj.onrender.com",
      "description": "Producción - Render"
    },
    {
      "url": "http://127.0.0.1:5000",
      "description": "Desarrollo local"
    }
  ],
...
  "paths": {
...
    "/api/admin/audit-logs": {
      "get": {
        "tags": [
          "Administration"
        ],
        "summary": "Consulta el historial de auditoría administrativa",
        "operationId": "listAdminAuditLogs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "description": "Búsqueda por nombre o correo del administrador.",
            "schema": {
              "type": "string",
              "maxLength": 120
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "description": "Filtra por operación registrada.",
            "schema": {
              "type": "string",
              "enum": [
                "product.created",
                "product.updated",
                "product.activated",
                "product.deactivated",
                "product.status_changed",
                "product.image_deleted",
                "product.deleted",
                "category.created",
                "category.updated",
                "category.deleted",
                "admin_user.created",
                "admin_user.updated",
                "admin_user.activated",
                "admin_user.deactivated",
                "order.confirmed",
                "order.cancelled",
                "order.status_changed"
              ]
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "description": "Tipo de entidad afectada.",
            "schema": {
              "type": "string",
              "enum": [
                "product",
                "category",
                "order",
                "admin_user"
              ]
            }
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "description": "ID exacto del registro afectado.",
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "description": "Fecha local de Colombia, inclusive, en formato AAAA-MM-DD.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "description": "Fecha local de Colombia, inclusive, en formato AAAA-MM-DD.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10000,
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Eventos de auditoría paginados.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuditLogListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Los filtros o parámetros de paginación no son válidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Se requiere autenticación administrativa.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT de acceso administrativo enviado como Authorization: Bearer <token>."
      }
    },
...
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "required": [
          "page",
          "per_page",
          "total",
          "pages",
          "has_next",
          "has_previous"
        ],
        "properties": {
          "page": {
            "type": "integer"
          },
          "per_page": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          },
          "has_next": {
            "type": "boolean"
          },
          "has_previous": {
            "type": "boolean"
          }
        }
      },
...
      "AdminAuditLog": {
        "type": "object",
        "required": [
          "id",
          "actor_id",
          "actor_name",
          "actor_email",
          "action",
          "entity_type",
          "entity_id",
          "details",
          "ip_address",
          "request_id",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "actor_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "actor_name": {
            "type": "string",
            "maxLength": 120
          },
          "actor_email": {
            "type": "string",
            "maxLength": 255
          },
          "action": {
            "type": "string",
            "enum": [
              "product.created",
              "product.updated",
              "product.activated",
              "product.deactivated",
              "product.status_changed",
              "product.image_deleted",
              "product.deleted",
              "category.created",
              "category.updated",
              "category.deleted",
              "admin_user.created",
              "admin_user.updated",
              "admin_user.activated",
              "admin_user.deactivated",
              "order.confirmed",
              "order.cancelled",
              "order.status_changed"
            ]
          },
          "entity_type": {
            "type": "string",
            "enum": [
              "product",
              "category",
              "order",
              "admin_user"
            ]
          },
          "entity_id": {
            "type": "string",
            "maxLength": 64,
            "nullable": true
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "ip_address": {
            "type": "string",
            "maxLength": 45,
            "nullable": true
          },
          "request_id": {
            "type": "string",
            "maxLength": 64,
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AdminAuditLogListResponse": {
        "type": "object",
        "required": [
          "success",
          "data",
          "pagination"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminAuditLog"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
    },
...
  }
}